Move write_volatile to the volatile module
